CBG-5784: split TestAddPendingLogs into its own file in package db
TestAddPendingLogs cannot move to db/changecachetest. It acquires
changeCache.lock and, while holding it, pushes directly onto the pendingLogs
heap and calls _pushRangeToPending and _addPendingLogs - underscore-prefixed
methods whose contract is that the caller already holds the lock. A held lock
cannot cross a package boundary, and an accessor that took the lock itself
would stop exercising the sequence-buffering path the test exists to pin down.

Moves it, with the four helpers only it uses, to change_cache_pending_test.go
so the remainder of change_cache_test.go can move as a whole file.
